Fix centre to build on Plan 9
Untested, but builds.
